+++ ts/features/messages/saga/handleLoadMessageData.ts
@@ -21,10 +21,7 @@ import { UIMessage, UIMessageDetails, UIMessageId } from "../types";
 import { serviceByIdPotSelector } from "../../services/details/store/reducers";
 import { loadServiceDetail } from "../../services/details/store/actions/details";
 import { messageDetailsByIdSelector } from "../store/reducers/detailsById";
-import {
-  extractContentFromMessageSources,
-  thirdPartyFromIdSelector
-} from "../store/reducers/thirdPartyById";
+import { thirdPartyFromIdSelector } from "../store/reducers/thirdPartyById";
 import { isLoadingOrUpdatingInbox } from "../store/reducers/allPaginated";
 import { TagEnum } from "../../../../definitions/backend/MessageCategoryPN";
 import { isPnEnabledSelector } from "../../../store/reducers/backendStatus/remoteConfig";
@@ -41,11 +38,9 @@ import {
 import { RemoteContentDetails } from "../../../../definitions/backend/RemoteContentDetails";
 import { MessageGetStatusFailurePhaseType } from "../store/reducers/messageGetStatus";
 import { ServicePublic } from "../../../../definitions/backend/ServicePublic";
-import {
-  ctaFromMessageCTA,
-  unsafeMessageCTAFromInput
-} from "../utils/messages";
+import { ctaFromMessageCTA, unsafeMessageCTAFromInput } from "../utils/ctas";
 import { isIoFIMSLink } from "../../../components/ui/Markdown/handlers/link";
+import { extractContentFromMessageSources } from "../utils";
 
 export function* handleLoadMessageData(
   action: ActionType<typeof getMessageDataAction.request>
